backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Login
Werkzeug
segno
